pyproject.toml
[tool.poetry]
name = "promptcards"
version = "0.1.0"
description = ""
authors = ["chkla <christopher.klamm@hotmail.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
frozen-flask = "^0.18"
requests = "^2.28.2"
pyyaml = "^6.0"
python-frontmatter = "^1.0.0"
markdown = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
